+++ docs/conf.py
@@ -21,7 +21,7 @@ project = "Pipeline Telemetry"
 year = "2021"
 author = "Maarten de Ruyter"
 copyright = "{0}, {1}".format(year, author)
-version = release = "0.4.4"
+version = release = "0.4.5"
 
 pygments_style = "trac"
 templates_path = ["."]